Use scoped packages, fix some typing issues
