+++ addons/@xterm/addon-attach/.npmignore
