+++ addons/@xterm/addon-attach/src/AttachAddon.api.ts
