+++ addons/@xterm/addon-fit/.gitignore
