+++ addons/@xterm/addon-fit/.npmignore
