+++ addons/@xterm/addon-fit/LICENSE
