+++ addons/@xterm/addon-fit/package.json
@@ -1,12 +1,12 @@
 {
-  "name": "xterm-addon-web-links",
+  "name": "@xterm/addon-fit",
   "version": "0.1.0-beta7",
   "author": {
     "name": "The xterm.js authors",
     "url": "https://xtermjs.org/"
   },
-  "main": "lib/WebLinksAddon.js",
-  "types": "typings/web-links.d.ts",
+  "main": "lib/FitAddon.js",
+  "types": "typings/fit.d.ts",
   "license": "MIT",
   "scripts": {
     "prepublishOnly": "../../node_modules/.bin/tsc -p src"
