+++ addons/@xterm/addon-fit/src/FitAddon.ts
