+++ addons/@xterm/addon-fit/src/tsconfig.json
@@ -14,6 +14,6 @@
   },
   "include": [
     "./**/*",
-    "../../../typings/xterm.d.ts"
+    "../../../../typings/xterm.d.ts"
   ]
 }
