+++ addons/@xterm/addon-search/.gitignore
