+++ addons/@xterm/addon-search/.npmignore
