+++ addons/@xterm/addon-search/LICENSE
