+++ addons/@xterm/addon-search/package.json
@@ -1,12 +1,12 @@
 {
-  "name": "xterm-addon-attach",
-  "version": "0.1.0-beta8",
+  "name": "@xterm/addon-search",
+  "version": "0.1.0-beta4",
   "author": {
     "name": "The xterm.js authors",
     "url": "https://xtermjs.org/"
   },
-  "main": "lib/AttachAddon.js",
-  "types": "typings/attach.d.ts",
+  "main": "lib/SearchAddon.js",
+  "types": "typings/search.d.ts",
   "license": "MIT",
   "scripts": {
     "prepublishOnly": "../../node_modules/.bin/tsc -p src"
