+++ addons/@xterm/addon-search/src/SearchAddon.ts
