+++ addons/@xterm/addon-search/src/tsconfig.json
@@ -4,16 +4,15 @@
     "target": "es5",
     "lib": [
       "dom",
-      "es2015"
+      "es6",
     ],
     "rootDir": ".",
     "outDir": "../lib",
     "sourceMap": true,
-    "removeComments": true,
-    "strict": true
+    "removeComments": true
   },
   "include": [
     "./**/*",
-    "../../../typings/xterm.d.ts"
+    "../../../../typings/xterm.d.ts"
   ]
 }
