+++ addons/@xterm/addon-web-links/.npmignore
