+++ addons/@xterm/addon-web-links/package.json
@@ -1,12 +1,12 @@
 {
-  "name": "xterm-addon-search",
-  "version": "0.1.0-beta4",
+  "name": "@xtermaddon-web-links",
+  "version": "0.1.0-beta7",
   "author": {
     "name": "The xterm.js authors",
     "url": "https://xtermjs.org/"
   },
-  "main": "lib/SearchAddon.js",
-  "types": "typings/search.d.ts",
+  "main": "lib/WebLinksAddon.js",
+  "types": "typings/web-links.d.ts",
   "license": "MIT",
   "scripts": {
     "prepublishOnly": "../../node_modules/.bin/tsc -p src"
