+++ addons/@xterm/addon-web-links/src/WebLinksAddon.api.ts
