+++ addons/@xterm/addon-web-links/src/tsconfig.json
@@ -1,7 +1,7 @@
 {
   "compilerOptions": {
     "module": "commonjs",
-    "target": "es2015",
+    "target": "es5",
     "lib": [
       "dom",
       "es2015"
@@ -14,6 +14,6 @@
   },
   "include": [
     "./**/*",
-    "../../../typings/xterm.d.ts"
+    "../../../../typings/xterm.d.ts"
   ]
 }
