+++ demo/tsconfig.json
@@ -6,10 +6,10 @@
     "sourceMap": true,
     "baseUrl": ".",
     "paths": {
-      "xterm-addon-attach": ["../addons/xterm-addon-attach"],
-      "xterm-addon-fit": ["../addons/xterm-addon-fit"],
-      "xterm-addon-search": ["../addons/xterm-addon-search"],
-      "xterm-addon-web-links": ["../addons/xterm-addon-web-links"]
+      "@xterm/addon-attach": ["../addons/@xterm/addon-attach"],
+      "@xterm/addon-fit": ["../addons/@xterm/addon-fit"],
+      "@xterm/addon-search": ["../addons/@xterm/addon-search"],
+      "@xterm/addon-web-links": ["../addons/@xterm/addon-web-links"]
     }
   },
   "include": [
