+++ tsconfig.all.json
@@ -3,9 +3,9 @@
   "include": [],
   "references": [
     { "path": "./src" },
-    { "path": "./addons/xterm-addon-attach/src" },
-    { "path": "./addons/xterm-addon-fit/src" },
-    { "path": "./addons/xterm-addon-search/src" },
-    { "path": "./addons/xterm-addon-web-links/src" }
+    { "path": "./addons/@xterm/addon-attach/src" },
+    { "path": "./addons/@xterm/addon-fit/src" },
+    { "path": "./addons/@xterm/addon-search/src" },
+    { "path": "./addons/@xterm/addon-web-links/src" }
   ]
 }
